Major Release
This is a bit unorthodox... but necessary to keep potential clients from seeing the system during migration while it isn't working.  Either way, on to the release summary.

This is a migration from html content, to ReactComponent markup.  With this transition, the majority of the UI has been written directly as Components... See Workflow.js

A few changes to the SceneDirector core will need to travel back to that repo.